Make COB property "RSCR" and "ISCR" into Commands instead
`RSCR` and `ISCR` were properties which required an equal sign and allowed only one value. For removal scripts (`RSCR`), that made sense, but for `ISCR`, it did not (for C1 COBs at least).
This commit shifts both properties to commands to improve install script includes, and to keep the syntax unified and also to support multiple install scripts.