Fix keywords losing color on block edit
Keywords where losing their coloring when editing a block because the keyword was temporarily made into an invalid loop terminator node.

This is due to the way the lexer is set up, which requires the full file to be lexed  all at once. Partial lexing leaves the tokens in an inconsistent state temporarily, an the highlighter does not update accordingly

To fix this issue (or rather mask it) we render the BAD_LOOP_TERMINATOR psi token as a keyword so it does not change highlighting