Allow collapsing long CAOS scripts for injection
Scripts were read one at a time as they were injected. This meant if a script was too long to inject, the injection would just stop, leaving game in an inconsistent state.

This commit allows collapsing scripts for injection, or aborting injection if any script is still too long, or if user does not want collapsed scripts.